Move JS files into `js/src` to distinguish between src and built
