+++ js/src/collections/shortcodes.js
