+++ js/src/utils/shortcode-view-constructor.js
