+++ js/src/views/edit-attribute-field.js
