+++ js/src/views/edit-shortcode-form.js
