+++ js/src/views/insert-shortcode-list-item.js
