+++ js/src/views/insert-shortcode-list.js
