+++ js/src/views/media-toolbar.js
