+++ js/src/views/search-shortcode.js
