+++ js/src/views/shortcode-preview.js
