+++ js/src/views/shortcode-ui.js
