+++ js/src/views/tabbed-view.js
